入力システムの不足機能追加

+++ System/Input/Format/InputFormatGamepadButton.cpp
@@ -1,6 +1,7 @@
 #include "InputFormatGamepadButton.h"
 #include <DxLib.h>
 #include "../System/Input/Gamepad.h"
+#include "../Utility/Vector3.h"
 
 void InputFormatGamepadButton::Init()
 {
@@ -29,5 +30,5 @@ bool InputFormatGamepadButton::CheckButtonState(const KeyCode::Button keyCode)
 
 Vector2 InputFormatGamepadButton::GetValue(const KeyCode::Button keyCode)
 {
-	return Vector2(static_cast<int>(CheckButtonState(keyCode)), 0.0f);
+	return Vector2(static_cast<float>(CheckButtonState(keyCode)), 0.0f);
 }
