+++ System/Input/Format/InputFormatKeyboard.cpp
@@ -1,6 +1,7 @@
 #include "InputFormatKeyboard.h"
 #include <DxLib.h>
 #include "../System/Input/Keyboard.h"
+#include "../Utility/Vector3.h"
 
 void InputFormatKeyboard::Init()
 {
@@ -131,5 +132,5 @@ bool InputFormatKeyboard::CheckButtonState(const KeyCode::Button keyCode)
 
 Vector2 InputFormatKeyboard::GetValue(const KeyCode::Button keyCode)
 {
-	return Vector2(static_cast<int>(CheckButtonState(keyCode)), 0.0f);
+	return Vector2(static_cast<float>(CheckButtonState(keyCode)), 0.0f);
 }
