+++ System/Input/Format/InputFormatMouseButton.cpp
@@ -20,5 +20,5 @@ bool InputFormatMouseButton::CheckButtonState(const KeyCode::Button keyCode)
 
 Vector2 InputFormatMouseButton::GetValue(const KeyCode::Button keyCode)
 {
-	return Vector2(static_cast<int>(CheckButtonState(keyCode)), 0.0f);
+	return Vector2(static_cast<float>(CheckButtonState(keyCode)), 0.0f);
 }
